Hibernate PostgreSQL json roundtrip

+++ src/main/java/io/pirkka/RestDemo/BookRepository.java
@@ -3,5 +3,5 @@ package io.pirkka.RestDemo;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BookRepository extends CrudRepository<Book, Long> {
-	Book findById(int id);
+	Book findById(long id);
 }
